requirements.txt
streamlit==1.39.0
openai==1.53.0
tiktoken==0.7.0
python-dotenv==1.0.1
pydantic==2.9.2
requests==2.32.3
beautifulsoup4==4.12.3
pandas==2.2.3
numpy==2.1.3
sentence-transformers==3.2.1
qdrant-client==1.9.0
anthropic==0.30.0